+++ src/reducers/root-reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { fake } from './fake-reducer';
-import { houseData } from './houseData-reducer';
+import houseData from './houseData-reducer';
 
 const rootReducer = combineReducers({
   fake,
